### 2024-11-12 #7450 @editorialbot
[PRE REVIEW]: Potnia: A Python library for the conversion of transliterated ancient texts to Unicode
**Submitting author:** @etour-unimelb (<a href="http://orcid.org/0000-0001-5212-1427">Emily Tour</a>)
**Repository:** https://github.com/AncientNLP/potnia.git
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.0
