### 2024-11-19 #7450 @rbturnbull
> @etour-unimelb – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week.

Dear @arfon - thanks for the message. Sorry we didn't respond until now. Yes, the code only has about 425 lines of Python code (in the main package plus tests). But a lot of the work in the project is in the YAML files which tell the package how to run and this is more than 2000 lines of code. We designed it so that YAML reduces the amount of code needed in Python and also makes it more accessible for domain experts to work with different languages. The PyConAU mentor team who did a pre-review of the code made a note about this in their recommendations to JOSS. Do you have their recommendation letter?
