### 2024-11-19 a43ed76
Update README.rst
:abc: adding JOSS badge
