### 2024-11-19 05deab2
Update README.rst
:abc: adding JOSS badge
