+++ README.rst
@@ -6,7 +6,7 @@ Potnia
 
 .. start-summary
 
-|pypi badge| |testing badge| |coverage badge| |docs badge| |git3moji badge| |black badge|
+|pypi badge| |testing badge| |coverage badge| |docs badge| |git3moji badge| |black badge| |JOSS badge|
 
 .. |pypi badge| image:: https://img.shields.io/pypi/v/potnia
     :target: https://pypi.org/project/potnia/
