### 2024-11-19 b327388
Update paper.bib
:abc: fixing broken DOI
