+++ paper.bib
@@ -58,7 +58,7 @@ journal = {Human IT: tidskrift för studier av IT ur ett humanvetenskapligt pers
 	volume = {24},
 	number = {1},
 	journal = {Zeitschrift für Antikes Christentum / Journal of Ancient Christianity},
-	doi = {doi:10.1515/zac-2020-0018},
+	doi = {10.1515/zac-2020-0018},
 	year = {2020},
 	lastchecked = {2024-09-26}
 }
