### 2024-11-19 #7450 @rbturnbull
I've fixed the broken DOI now.
