### 2024-11-19 b32137e
:zap: :cop: Update to LA mappings to accommodate new transliterations. Addition of test cases. Slight modifications to wording in LB mappings.
