+++ potnia/data/linear_a.yaml
@@ -1,488 +1,481 @@
 mappings:
   ############################################################
-  ###### Syllabograms Common to Linear A and Linear B ########
+  ######  Signs Common to Linear A and Linear B ########
   ############################################################
-  a: 𐀀
-  e: 𐀁
-  i: 𐀂
-  o: 𐀃
-  u: 𐀄
-  da: 𐀅
-  de: 𐀆
-  di: 𐀇
-  do: 𐀈
-  du: 𐀉
-  ja: 𐀊
-  je: 𐀋
-  jo: 𐀍
-  ju: 𐀎
-  ka: 𐀏
-  ke: 𐀐
-  ki: 𐀑
-  ko: 𐀒
-  ku: 𐀓
-  ma: 𐀔
-  me: 𐀕
-  mi: 𐀖
-  mo: 𐀗
-  mu: 𐀘
-  na: 𐀙
-  ne: 𐀚
-  ni: 𐀛
-  "no": 𐀜
-  nu: 𐀝
-  nwa: 𐁅
-  pa: 𐀞
-  pe: 𐀟
-  pi: 𐀠
-  po: 𐀡
-  pu: 𐀢
-  pu₂: 𐁆
-  pu2: 𐁆
-  qa: 𐀣
-  qe: 𐀤
-  qi: 𐀥 #𐘏?
-  qif: 𐘐
-  qo: 𐀦
-  ra: 𐀨
-  ra₂: 𐁈
-  ra2: 𐁈
-  re: 𐀩
-  ri: 𐀪
-  ro: 𐀫
-  ru: 𐀬
-  sa: 𐀭
-  se: 𐀮
-  si: 𐀯
-  so: 𐀰
-  su: 𐀱
-  ta: 𐀲
-  ta₂: 𐁋
-  ta2: 𐁋
-  te: 𐀳
-  ti: 𐀴
-  to: 𐀵
-  tu: 𐀶
-  wa: 𐀷
-  we: 𐀸
-  wi: 𐀹
-  wo: 𐀺
-  za: 𐀼
-  ze: 𐀽
-  zo: 𐀿
 
-  "*22": 𐁒
-  "*22f": 𐘓
-  "*34": 𐁓
-  "*47": 𐁔
-  "*49": 𐁕
-  "*56": 𐘰
-  "*72": "?" #FIND # 𐘽 ?
-  "*79": 𐙀
+  # SYLLABOGRAMS (LOWERCASE)
+  a: 𐘇
+  e: 𐘡
+  i: 𐘚
+  o: 𐘵
+  u: 𐘉
+  au: 𐙄 #also SUS
+  da: 𐘀
+  de: 𐘦
+  di: 𐘆
+  du: 𐘬
+  ja: 𐘱
+  je: 𐘧
+  ju: 𐘶
+  ka: 𐘾
+  ke: 𐘥
+  ki: 𐘸
+  ko: 𐘺
+  ku: 𐙂
+  ma: 𐙁
+  me: 𐘋
+  mi: 𐘻
+  mu: 𐘕
+  na: 𐘅
+  ne: 𐘗
+  ni: 𐘝 #also FIC
+  nu: 𐘯
+  nwa: 𐘩
+  pa: 𐘂
+  pi: 𐘢
+  po: 𐘊
+  pu: 𐘫
+  pu₂: 𐘜
+  qa: 𐘌
+  qe: 𐘿
+  qi: 𐘏 #also OVIS
+  qif: 𐘐 #also OVISf
+  ra: 𐘴
+  ra₂: 𐘽
+  ra2: 𐘽
+  re: 𐘙
+  ri: 𐀪 𐘭
[…]
