+++ tests/expected/linear_a_unicode_regularized.yaml
@@ -1,9 +1,28 @@
-# Test that syllabograms and logograms are correctly converted to Unicode, that spaces work as word separators, and that hyphens act to join syllables within a word
-"ka-di AB131/VINa 3 ku-ro AB131/VINa 78 A594 17" : "𐀏𐀇 𐙍 3 𐀓𐀫 𐙍 78 𐛿 17" # https://sigla.phis.me/document/ZA%2015b/
+# Test that signs are correctly converted to Unicode, that spaces work as word separators, and that hyphens act to join syllables within a word.
+# Test that the special square character is tokenized, and converted to a wildcard character ('%') in regularized output.
+# Test that '—' is tokenized, but not included in regularized output.
+#HT49b
+    "𐝫  KU  I-KI TA  —  𐝫  KA  𐝫" : "% 𐙂 𐘚𐘸 𐘳 % 𐘾 %" #HT49b
 
-# Test that ']' and '[' are correctly converted to wildcard characters ('%')
-# Test that '[?]'' is correctly converted to a wildcard character ('%')
-"]pa-ri-de ]a-si-*118 ]ku-ka-[?][" : "%𐀞𐀪𐀆 %𐀀𐀯𐙈 %𐀓𐀏%%" # https://sigla.phis.me/document/KH%2099/index-7.html
+# Test that '≈' is tokenized, but not included in regularized output.
+    "165 ≈ ¹⁄₆  KI-RO" : "165 𐝅 𐘸𐘁" #HT93b
 
-# Test that '[unclassified]' is correctly converted to a wildcard character ('%')
-"[unclassified] *180 A339 *100" : "% 𐙒 𐙿 𐙇" # https://sigla.phis.me/document/PH%2012a/
+# Test that 'double mina' is tokenized, so that it can be picked up by the correct YAML mapping.
+# Test that '?' is tokenized, so that it can be picked up by the correct YAML mapping.
+# Test that '𐄁' is tokenized, but not included in regularized output
+    "DA-RE-KU 𐄁  KU-*321 𐄁 GRA+PA 2 double mina  GRA 2 double mina  OLE+KI ?  OLIV ¹⁄₄  𐝫" : "𐘀𐘙𐙂 𐄁 𐙂𐙭 𐄁 𐛭 2 𐝂𐝂 𐙉 2 𐝂𐝂 𐜓 𐝁𐝁 𐙋 𐝃 %" #ZA18a
+
+# Test that '+' signs concatenate the signs on either side into a single token. This includes when the '+' sign is followed by a '[?]'' or '[ ]'.
+# There can be multiple '+' signs in a single token.
+# Make sure not to break legitimate cases where there is a '?' by itself, or a '[' that's not part of this pattern (see test cases immediately above and below).
[…]
