### 2024-11-20 e6e5c17
:cop: Minor change to LA test case descriptions
