### 2024-11-26 dcc32c4
Fixing Linear A issues
