+++ potnia/data/linear_a.yaml
@@ -308,8 +308,8 @@ mappings:
   "*371": 𐚟
 
   # SPECIAL CHARACTERS
-    "double mina": 𐝂𐝂
-    "?": 𐝁𐝁
+  "double mina": 𐝂𐝂
+  "?": 𐝁𐝁
 
   # FRACTIONS
   "¹⁄₆": 𐝀
@@ -437,6 +437,7 @@ mappings:
   "QA2+[?]+RE": 𐚹
   "QA2+[?]+PU+RE": 𐚼
   "QA2+[?]+RE+PU": 𐚺
+  
   "QE+GRA+PA": 𐛟
   "QI+SI": 𐚽
   "RA+KA": 𐛔
@@ -498,73 +499,20 @@ patterns_to_ignore:
 
 
 regularization:
-  - ['\\[\\?\\]', '%']
-  - ['\\[unclassified\\]', '%']  # Maps `[unclassified]` to `%`
-  - ['[?]','%']
-  - ['\\|', '']
-  - [':', '']
-  - ['r\\.', '']
-  - ['\\[•~\\]', '']
-  - ['⌜', '']
-  - ['⌝', '']
-  - ['mutila', '']
-  - ['\\[?•~•~•~•\\]?', '%%%%']
-  - ['\\[?•~•~•\\]?', '%%%']
-  - ['\\[?•~•~\\]?', '%%']
-  - ['\\[?•~•\\]?', '%%']  # Corrected this line
-  - ['\\<|\\>', '']
-  - ['\\[ \\]', '[ ]']
-  - ['ro2', '𐁊']
-  - ['vestigia', '%']
-  - ['\\bqs\\b', '%']
-  - ['vest\\s*\\.', '%']
-  - ['\\[•\\]', '%']
-  - ['supra sigillum|CMS \\w+\\d+[A-Z]* \\d+', '']
-  - ['reliqua pars sine regulis', '']
-  - ['[αβγ]', '']
-  - ['v\\.→', '']
-  - ['v\\.↓', '']
-  - ['v\\.', '']
-  - ['\\b(vacat|sup. mut.|inf. mut.|deest|X|fragmentum A|fragmentum B|graffito|angustum|prior pars sine regulis|fragmentum C|fragmentum D|fragmentum separatum|α|β|γ|δ|sigillum|)\\b', '']  # Corrected this line
-  - ['\\b(x|m|f)\\b', '']
+  - ['—','']
+  - ['≈','']
   - ['[\\[\\]]', '%']
-  - ['=[^ ]*', '']
-  - ['•', '%']
-  - ['●', '']
-  - ['dex.', '']
-  - ['sin.', '']
+  - ['\U0001076b','%']
 
 tokenization:
-  - ['\u00a0', ' ']  # Replace non-breaking space with regular space
-  - ['\u0323', '']   # Remove specific character (e.g., dot below)
-  - ['</em>', '']    # Remove HTML closing emphasis tag
-  - ['<em>', '']     # Remove HTML opening emphasis tag
-  - ['\|([^|]+)\|', '|\1|']  # Special handling to ensure pipes are treated as separate tokens
-  - ['ME<±RI>', 'ME±RI']  # Handle specific compound tokens like 'ME<±RI>'
-  - ['--', '-']  # Normalize the text by replacing double dashes with a single dash
-  - ['\b(EQU|SUS|OVIS|BOS|CAP)\s+(x|m|f)\b', '\1\2']  # Combine animal ideograms followed by 'x', 'm', or 'f' without space
-  - ['⌜', ' ⌜ ']  # Explicit tokenization for half brackets
-  - ['⌝', ' ⌝ ']
[…]
