+++ tests/expected/linear_a_unicode.yaml
@@ -1,11 +1,0 @@
-# # Test that syllabograms and logograms are correctly converted to Unicode, that spaces work as word separators, and that hyphens act to join syllables within a word
-# "ka-di AB131/VINa 3 ku-ro AB131/VINa 78 A594 17" : "𐀏𐀇 𐙍 3 𐀓𐀫 𐙍 78 𐛿 17" # https://sigla.phis.me/document/ZA%2015b/
-
-# # Test that ']' and '[' are correctly tokenized as individual tokens
-# # Test that '[?]'' is correctly converted to a wildcard character ('%')
-# "]pa-ri-de ]a-si-*118 ]ku-ka-[?][" : "]𐀞𐀪𐀆 ]𐀀𐀯𐙈 ]𐀓𐀏[?][" # https://sigla.phis.me/document/KH%2099/index-7.html
-
-# # Test that '[unclassified]' is correctly tokenized as a single token
-# "[unclassified] *180 A339 *100" : "[unclassified] 𐙒 𐙿 𐙇" # https://sigla.phis.me/document/PH%2012a/
-
-# "ka-di[unclassified] *180 A339 *100" : "𐀏𐀇[unclassified] 𐙒 𐙿 𐙇"
