### 2024-11-26 7677bce
:cop: One more LA test case for LAXYZ data
