### 2024-11-26 5b58601
:zap: Removed incorrect LA mappings
