+++ potnia/data/linear_a.yaml
@@ -476,11 +476,6 @@ mappings:
   "VIR+*313c": 𐛫
   "WI+ZE": 𐛇
 
-  # DAMAGED COMPOUND SIGNS
-  "HIDE+": 𐙒+
-  "VIR+": 𐙇+
-  "*304+": 𐙘+
-
   # ADDITIONAL SIGNS
   "*800": 𐝠
   "*801": 𐝡
