### 2024-11-26 5543a9e
Fixing complex symbols mapping linear a
