### 2024-11-26 53d998c
:zap: Added one more LA mapping
