### 2024-11-28 0029a39
:zap: Update to a few Linear B sign descriptions
