+++ potnia/data/linear_b.yaml
@@ -8,7 +8,6 @@ mappings:
   i: 𐀂
   o: 𐀃
   u: 𐀄
-  au: 𐁂
   da: 𐀅
   de: 𐀆
   di: 𐀇
@@ -33,21 +32,16 @@ mappings:
   ni: 𐀛
   "no": 𐀜
   nu: 𐀝
-  nwa: 𐁅
   pa: 𐀞
   pe: 𐀟
   pi: 𐀠
   po: 𐀡
   pu: 𐀢
-  pu₂: 𐁆
-  pu2: 𐁆
   qa: 𐀣
   qe: 𐀤
   qi: 𐀥
   qo: 𐀦
   ra: 𐀨
-  ra₂: 𐁈
-  ra2: 𐁈
   re: 𐀩
   ri: 𐀪
   ro: 𐀫
@@ -58,13 +52,10 @@ mappings:
   so: 𐀰
   su: 𐀱
   ta: 𐀲
-  ta₂: 𐁋
-  ta2: 𐁋
   te: 𐀳
   ti: 𐀴
   to: 𐀵
   tu: 𐀶
-  twe: 𐁌
   wa: 𐀷
   we: 𐀸
   wi: 𐀹
@@ -73,12 +64,23 @@ mappings:
   ze: 𐀽
   zo: 𐀿
 
+# subsyllabary
+  au: 𐁂
+  nwa: 𐁅
+  pu₂: 𐁆
+  pu2: 𐁆
+  ra₂: 𐁈
+  ra2: 𐁈
+  ta₂: 𐁋
+  ta2: 𐁋
+  twe: 𐁌
+
   # LB SOUND VALUES UNKNOWN
-  "*22": 𐁒
-  "*34": 𐁓
-  "*47": 𐁔
+  "*22": 𐁒 # pi₂?
+  "*34": 𐁓 # a5?
+  "*47": 𐁔 #silver?
   "*49": 𐁕
-  "*86": 𐁜
+  "*86": 𐁜 # dwa?
 
   ############################################################
   ######      Syllabograms Unique to Linear B         ########
@@ -103,15 +105,15 @@ mappings:
   "*19": 𐁑
   # *35 missing unicode sign?
   "*63": 𐁗
-  "*83": 𐁛
+  "*83": 𐁛 #nwe?
   "*89": 𐁝
 
   # Doubtful sound values
   "*56": 𐁖 # potential sound value 'pa₃'
   "*64": 𐁘 # potential sound value 'swi'
   "*65": 𐀎 # potential sound value 'ju'
-  "*79": 𐁙 # potential sound value 'zu'
-  "*82": 𐁚 # potential sound value 'swa'
+  "*79": 𐁙 # potential sound value 'zu' or 'wo2'
+  "*82": 𐁚 # potential sound value 'swa' or 'twa'
 
   ############################################################
   #####                  Logograms                     #######
@@ -163,8 +165,8 @@ mappings:
   AROM: 𐂑 # spices
   CYP: 𐂒 # cyperus
   PYC: 𐂒 # cyperus
-  KA+PO: 𐂓 # fruit?
-  KA±PO: 𐂓 # fruit?
+  KA+PO: 𐂓 # fruit? bark?
+  KA±PO: 𐂓 # fruit? bark?
   KA+NA+KO: 𐂔 # saffron
   KA±NA±KO: 𐂔 # saffron
   KANAKO: 𐂔 # saffron
@@ -174,63 +176,57 @@ mappings:
   # LIQUID COMMODITIES
   OLE: 𐂕 # olive oil
   VIN: 𐂖 # wine
-  "*132": 𐂗
[…]
