### 2025-01-18 #7450 @arfon
:wave: @afrubin @claresloggett – I believe you both did the PyCon AU review, would you also be willing to complete a JOSS review here for this submission?
