### 2025-01-19 #7450 @rbturnbull
> 👋 [@afrubin](https://github.com/afrubin) [@claresloggett](https://github.com/claresloggett) – I believe you both did the PyCon AU review, would you also be willing to complete a JOSS review here for this submission?

I think @wytamma was also involved in the PyCon AU review.
