### 2025-01-19 #7450 @afrubin
> 👋 [@afrubin](https://github.com/afrubin) [@claresloggett](https://github.com/claresloggett) – I believe you both did the PyCon AU review, would you also be willing to complete a JOSS review here for this submission?

I'm out of town currently, but very happy to take on this review when I get back in a little over a week.
