### 2025-01-20 #7450 @arfon
> I think @Wytamma was also involved in the PyCon AU review.

Would you also be willing to complete a JOSS review here for this submission?
