### 2025-01-26 #7450 @arfon
@jjmccollum @mikahama – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _Potnia: A Python library for the conversion of transliterated ancient texts to Unicode_

Based on your experience, and past papers in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
