### 2025-02-11 9492704
:tv: adding tests for py3.12 in github actions and restricting to <3.13
