+++ pyproject.toml
@@ -17,13 +17,14 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = ">=3.10,<3.13"
 pyyaml = "^6.0.1"
 typer = "^0.12.5"
 guigaga = ">=0.0.5"
 numpy = "<2"
 pybtex = ">=0.24.0"
 pybtexnbib = ">=0.1.1"
+setuptools = "^75.8.0"
 
 
 [tool.poetry.group.dev.dependencies]
