### 2025-02-17 #7725 @claresloggett
@arfon Can you give me any pointers on where editorialbot gets data for the software report https://github.com/openjournals/joss-reviews/issues/7725#issuecomment-2615582696 ?

The commit count on the github "Insights" page looks quite different. It excludes the 58 commits attributed by editorialbot to "Emily Tour" (the lead author) and only shows the handful of commits from etour-unimelb; presumably the same person. The Insights page also does not list "Kabir Manandhar Shrestha" at all.

I am assuming editorialbot's numbers are right, as they are larger, and they match the author list. Is the Insights page just not reliable?
