+++ tests/expected/hittite_unicode.yaml
@@ -1,10 +1,10 @@
-"ḫe]-en-ku-un šu-me-eš ma-ni-ia-aḫ-ḫi-eš-ke-et-tén" : "𒄭]𒂗𒆪𒌦 𒋗𒈨𒐁 𒈠𒉌𒅀𒄴𒄭𒐁𒆠𒀉𒁷"
+"ḫe]-en-ku-un šu-me-eš ma-ni-ia-aḫ-ḫi-eš-ke-et-tén" : "𒄭𒂗𒆪𒌦 𒋗𒈨𒐁 𒈠𒉌𒅀𒄴𒄭𒐁𒆠𒀉𒁷"
 
 "a-ši KÁ-aš ku-iš pa-it nu DINGIR (MEŠ)  2 ḪUR.SAG (MEŠ)" : "𒀀𒅆 𒆍𒀸 𒆪𒅖 𒉺𒀉 𒉡 𒀭 𒌍  2 𒄯𒊕 𒌍"
 
-"ki‑nu‑na‑an an‑ni‑iš‑ke‑mi ku‑u‑un" : "𒆠𒉡𒈾𒀭 𒀭𒉌𒅖𒆠𒈪 𒆪𒌋𒌦"
+"ki-nu-na-an an-ni-iš-ke-mi ku-u-un" : "𒆠𒉡𒈾𒀭 𒀭𒉌𒅖𒆠𒈪 𒆪𒌋𒌦"
 
-"[UD].KAM‑an": "𒌓𒄰𒀭"
+"[UD].KAM-an": "𒌓𒄰𒀭"
 
 "12": "𒌋𒈫"
 
@@ -19,6 +19,6 @@
 
 "UZU.SA": "𒍜𒊓"
 
-"ḫa‑⸢an⸣‑da‑⸢nu⸣‑un": "𒄩𒀭𒁕𒉡𒌦"
+"ḫa-⸢an⸣-da-⸢nu⸣-un": "𒄩𒀭𒁕𒉡𒌦"
 
 "ŠA": "𒊭"
