### 2025-02-24 #7725 @afrubin
@etour-unimelb I've completed my review and opened a few minor issues in the repository for you to consider.

@arfon I don't think any of these should block publication. They're all just nice-to-haves.
