### 2025-02-25 0187e29
:tv: ignoring DeprecationWarning in dependencies during pytest (#9)
