+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "potnia"
-version = "0.3.2"
+version = "0.3.3"
 description = "Potnia is an open-source Python library designed to convert Romanized transliterations of ancient texts into Unicode representations of ther respective native scripts.."
 authors = ["Emily Tour <esmtour@gmail.com>", "Kabir Manandhar Shrestha", "Robert Turnbull"]
 license = "Apache-2.0"
@@ -41,6 +41,8 @@ myst-parser = "^3.0.1"
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
+[tool.pytest.ini_options]
+filterwarnings = ["ignore::DeprecationWarning"]
 
 [tool.poetry.scripts]
 potnia = "potnia.main:app"
