### 2025-02-25 2598df4
docs: add glyph display guide and reference in installation section
