+++ docs/glyphs.rst
@@ -1,0 +1,9 @@
+Viewing Ancient Glyphs
+======================
+
+If the glyphs for Linear A, Linear B, or other ancient scripts do not display correctly in your terminal, code editor, web browser, or Jupyter notebook, it is likely due to missing fonts. Install the following Noto Sans fonts from Google Fonts for proper rendering:
+
+- `Noto Sans Linear A <https://fonts.google.com/noto/specimen/Noto+Sans+Linear+A>`_
+- `Noto Sans Linear B <https://fonts.google.com/noto/specimen/Noto+Sans+Linear+B>`_
+
+Ensure your editor, terminal, or application is set to use these fonts. If you are browsing the repository online, **Firefox** typically offers better glyph support compared to **Google Chrome** or **Microsoft Edge**.
