+++ README.rst
@@ -66,6 +66,8 @@ To install the latest version from the repository, you can use this command:
 
     pip install git+https://github.com/AncientNLP/potnia.git
     
+Note: For proper display of ancient script glyphs, please refer to the :doc:`glyphs` section.
+
 Usage
 ====================
 
