### 2025-02-25 33f8493
Fixing reference to adding glyphs doc in readme
