### 2025-02-25 bc8f7d8
Adding fonts sections to the docs
