### 2025-02-25 23f42e0
Fonts section reference to readme.rst
