### 2025-02-26 #7725 @rbturnbull
Hi @claresloggett - I think you're right that editorialbot's numbers from cloc are a more reliable indicator. The insights page has 'Contributions per week to main' at the top so it might be that Kabir's commits were more on branches that were later merged than on main directly. I can assure you that both he and Emily made massive contributions to the code base as we developed this project together. All three of us are listed as `equal-contrib: true` on the paper.
