### 2025-02-26 fbd84bd
:abc: adding more instructions about installing from poetry in the README and quickstart. See issue #8
