+++ mkdocs.sh
@@ -1,1 +1,2 @@
 sphinx-build -b html docs docshtml -E -a
+echo docshtml/index.html
