### 2025-02-27 e8aba6b
Enhance  function to handle uppercase/lowercase mappings
- Updated  to first check for the original token in the YAML mapping.
- If no match is found, it attempts to find the corresponding opposite case (lowercase → uppercase or vice versa).
- Ensures correct mapping for Hittite symbols that exist only in one case.
- Preserves existing functionality while improving robustness for mixed-case transliterations.
