### 2025-03-01 AncientNLP/potnia#10 @claresloggett
Empty test yaml file
Great work on potnia! This really looks like a solid and useful tool.

When I run tests, one test is being skipped, and it looks like it's because the file `tests/expected/linear_a_transliteration.yaml` is empty, i.e. there are no test cases for the parametrised test. Is this a known issue?

Related to review https://github.com/openjournals/joss-reviews/issues/7725
