### 2025-03-01 AncientNLP/potnia#10 @rbturnbull
Thanks @claresloggett - I'll add some tests there now.
