### 2025-03-01 50c1252
:abc: adding funding acknowledgement
