+++ tests/expected/linear_a_transliteration.yaml
@@ -1,0 +1,3 @@
+"𐙒 𐙿 𐙇" : "*180 A339 *100"  # https://sigla.phis.me/document/PH%2012a/
+"𐀏𐀇 𐙒 𐙿 𐙇" : "ka-di *180 A339 *100"
+"𐀏𐀇 𐙍 3 𐀓𐀫 𐙍 78 𐛿 17" : "ka-di *131a 3 ku-ro *131a 78 A594 17"  # https://sigla.phis.me/document/ZA%2015b/. Uses *131a instead of AB131/VINa
