### 2025-03-01 AncientNLP/potnia#10 @rbturnbull
Thanks @claresloggett  - I've added the tests now.
