### 2025-03-01 c9d7fb3
:zap: supporting -h for help info (see issue #11)
