### 2025-03-14 531bd19
Add Lycian support in Potnia: YAML mappings and test cases
