+++ potnia/main.py
@@ -5,6 +5,8 @@ from potnia import linear_b as linear_b_script
 from potnia import hittite as hittite_script
 from potnia import arabic as arabic_script
 from potnia import lydian as lydian_script
+from potnia import lycian as lycian_script
+
 # from potnia import luwian as luwian_script
 # from potnia import akkadian as akkadian_script
 
@@ -74,6 +76,13 @@ def lydian(text: list[str]=TEXT_ARGUMENT, regularize:bool=REGULARIZATION_DEFAULT
         text = " ".join(text)
     print(lydian_script(text, regularize=regularize))
 
+
+@app.command()
+def lycian(text: list[str]=TEXT_ARGUMENT, regularize:bool=REGULARIZATION_DEFAULT):
+    """ Converts a Lydian text to Unicode. """
+    if isinstance(text, list):
+        text = " ".join(text)
+    print(lycian_script(text, regularize=regularize))
 
 @app.command()
 def bibtex():
