+++ potnia/__init__.py
@@ -5,6 +5,7 @@ from .scripts.hittite import hittite, Hittite
 from .scripts.arabic import arabic, Arabic
 from .scripts.lydian import lydian, Lydian
 from .scripts.lycian import lycian, Lycian
+from .scripts.luwian_cuneiform import luwian_cuneiform,Luwian_Cuneiform
 
 
 
