+++ tests/expected/luwian_cuneiform_unicode.yaml
@@ -1,0 +1,2 @@
+"tar-na-aš ŠU.GI-iš ti-ya-a": "𒋻𒈾𒀸 𒋗𒃻𒅖 𒋾𒅀𒀀"
+"wa-aš-ša ta-pa-ru-uš": "𒉿𒀸𒊭 𒋫𒉺𒊒𒍑"
