+++ potnia/data/luwian_cuneiform.yaml
@@ -1,0 +1,415 @@
+mappings:
+  # V syllabograms
+  "a": 𒀀
+  "e": 𒂊
+  "i": 𒄿
+  "u": 𒌋
+  "ú": 𒌑
+  "ia": 𒅀
+
+  # CV syllabograms
+  "ba": 𒁀
+  "be": 𒁁
+  "bi": 𒁉
+  "bu": 𒁍
+
+  "pa": 𒉺
+  "pé": 𒁉
+  "pí": 𒁉
+  "pu": 𒁍
+
+  "da": 𒁕
+  "de": 𒁲
+  "di": 𒁲
+  "du": 𒁺
+
+  "ta": 𒋫
+  "te": 𒋼
+  "ti": 𒋾
+  "tu": 𒌅
+
+  "ga": 𒂵
+  "ge": 𒄀
+  "gi": 𒄀
+  "gu": 𒄖
+
+  "ka": 𒅗
+  "ke": 𒆠
+  "ki": 𒆠
+  "ku": 𒆪
+
+  "ḫa": 𒄩
+  "ḫe": 𒄭
+  "ḫé": 𒃶
+  "ḫi": 𒄭
+  "ḫu": 𒄷
+
+  "la": 𒆷
+  "le": 𒇷
+  "li": 𒇷
+  "lu": 𒇻
+
+  "ma": 𒈠
+  "me": 𒈨
+  "mé": 𒈪
+  "mi": 𒈪
+  "mu": 𒈬
+
+  "na": 𒈾
+  "ne": 𒉈
+  "né": 𒉌
+  "ni": 𒉌
+  "nu": 𒉡
+
+  "ra": 𒊏
+  "re": 𒊑
+  "ri": 𒊑
+  "ru": 𒊒
+
+  "ša": 𒊭
+  "še": 𒊺
+  "ši": 𒅆
+  "šu": 𒋗
+  "šú": 𒋙
+
+  "wa": 𒉿
+  "wi5": 𒃾
+
+  "ya": 𒅀
+
+  "za": 𒍝
+  "ze": 𒍣
+  "zé": 𒍢
+  "zi": 𒍣
+  "zu": 𒍪
+
+  # VC syllabograms
+  "ab": 𒀊
+  "ap": 𒀊
+  "eb": 𒅁
+  "ep": 𒅁
+  "ib": 𒅁
+  "ip": 𒅁
+  "ub": 𒌒 
+  "up": 𒌒
+
+  "ad": 𒀜
+  "at": 𒀜
+  "ed": 𒀉
+  "et": 𒀉
+  "id": 𒀉
+  "it": 𒀉
+  "ud": 𒌓 
+  "ut": 𒌓
+
+  "ag": 𒀝 
+  "ak": 𒀝
+  "eg": 𒅅
+  "ek": 𒅅
+  "ig": 𒅅
+  "ik": 𒅅
+  "ug": 𒊌
+  "uk": 𒊌
+
+  "aḫ": 𒄴
+  "eḫ": 𒄴
+  "iḫ": 𒄴
+  "uḫ": 𒄴
+
+  "al": 𒀠
+  "el": 𒂖
+  "il": 𒅋
+  "ul": 𒌌 # Check
+
+  "am": 𒄠
+  "em": 𒅎
+  "im": 𒅎
+  "um": 𒌝
+
+  "an": 𒀭
+  "en": 𒂗
+  "in": 𒅔
+  "un": 𒌦
+
+  "ar": 𒅈
+  "er": 𒅕
+  "ir": 𒅕
+  "ur": 𒌨
+  "úr": 𒌫
+
+  "aš": 𒀸
+  "eš": 𒌍
[…]
