+++ tests/expected/luwian_cuneiform_unicode.yaml
@@ -1,2 +1,2 @@
-"tar-na-aš ŠU.GI-iš ti-ya-a": "𒋻𒈾𒀸 𒋗𒃻𒅖 𒋾𒅀𒀀"
-"wa-aš-ša ta-pa-ru-uš": "𒉿𒀸𒊭 𒋫𒉺𒊒𒍑"
+"tar-na-aš ŠA-iš ti-ya-a": "𔑕𔑞𔑙𔐀𔑩 𔑤𔐁𔑩 𔑖𔓱𔐀"
+"wa-aš-ša ta-pa-ru-uš": "𔑣𔐀𔑩𔑤 𔑕𔑒𔑟𔐂𔑩"
