+++ tests/expected/luwian_hieroglyhpic_unicode.yaml
@@ -1,0 +1,2 @@
+"tar-na-aš ŠA-iš ti-ya-a": "𔑕𔑞𔑙𔐀𔑩 𔑤𔐁𔑩 𔑖𔓱𔐀"
+"wa-aš-ša ta-pa-ru-uš": "𔑣𔐀𔑩𔑤 𔑕𔑒𔑟𔐂𔑩"
