### 2025-04-03 4c5cd9d
Final copy-edit updates for paper.md prior to publication
