### 2025-04-03 fb4d37c
:tv: Bumping to v0.4.0
